vault hsm 버전 이미지

+++ Dockerfile
@@ -1,40 +1,101 @@
-FROM ubuntu:20.04
+FROM registry.access.redhat.com/ubi8/ubi-minimal:8.6
 
-# This is the release of Vault to pull in.
-ARG VAULT_VERSION=1.10.5+ent.hsm
+LABEL maintainer="HashiCorp"
+ARG VAULT_VERSION=1.11.0+ent.hsm
 
-# Create a vault user and group first so the IDs get set the same way,
-# even as the rest of this may change over time.
-RUN groupadd vault && \
-    useradd -g vault vault
+# Additional metadata labels used by container registries, platforms
+# and certification scanners.
+LABEL name="Vault" \
+      maintainer="Vault Team <vault@hashicorp.com>" \
+      vendor="HashiCorp" \
+      version=$VAULT_VERSION \
+      release=$VAULT_VERSION \
+      summary="Vault is a tool for securely accessing secrets." \
+      description="Vault is a tool for securely accessing secrets. A secret is anything that you want to tightly control access to, such as API keys, passwords, certificates, and more. Vault provides a unified interface to any secret, while providing tight access control and recording a detailed audit log."
+
+# Copy license for Red Hat certification.
+# COPY LICENSE /licenses/mozilla.txt
 
 # Set up certificates, our base tools, and Vault.
-RUN apt-get -y update && \
-    apt-get -y upgrade && \
-    apt-get install -y wget ca-certificates unzip && \
+RUN set -eux; \
+    microdnf install -y ca-certificates gnupg openssl libcap tzdata wget unzip procps shadow-utils util-linux && \
+    VAULT_GPGKEY=C874011F0AB405110D02105534365D9472D7468F; \
+    found=''; \
+    for server in \
+        hkps://keys.openpgp.org \
+        hkps://keyserver.ubuntu.com \
+        hkps://pgp.mit.edu \
+    ; do \
+        echo "Fetching GPG key $VAULT_GPGKEY from $server"; \
+        gpg --batch --keyserver "$server" --recv-keys "$VAULT_GPGKEY" && found=yes && break; \
+    done; \
+    test -z "$found" && echo >&2 "error: failed to fetch GPG key $VAULT_GPGKEY" && exit 1; \
     mkdir -p /tmp/build && \
     cd /tmp/build && \
     wget https://releases.hashicorp.com/vault/${VAULT_VERSION}/vault_${VAULT_VERSION}_linux_amd64.zip && \
     wget https://releases.hashicorp.com/vault/${VAULT_VERSION}/vault_${VAULT_VERSION}_SHA256SUMS && \
     wget https://releases.hashicorp.com/vault/${VAULT_VERSION}/vault_${VAULT_VERSION}_SHA256SUMS.sig && \
+    gpg --batch --verify vault_${VAULT_VERSION}_SHA256SUMS.sig vault_${VAULT_VERSION}_SHA256SUMS && \
+    grep vault_${VAULT_VERSION}_linux_amd64.zip vault_${VAULT_VERSION}_SHA256SUMS | sha256sum -c && \
     unzip -d /tmp/build vault_${VAULT_VERSION}_linux_amd64.zip && \
     cp /tmp/build/vault /bin/vault && \
     if [ -f /tmp/build/EULA.txt ]; then mkdir -p /usr/share/doc/vault; mv /tmp/build/EULA.txt /usr/share/doc/vault/EULA.txt; fi && \
     if [ -f /tmp/build/TermsOfEvaluation.txt ]; then mkdir -p /usr/share/doc/vault; mv /tmp/build/TermsOfEvaluation.txt /usr/share/doc/vault/TermsOfEvaluation.txt; fi && \
     cd /tmp && \
     rm -rf /tmp/build && \
-    apt-get purge -y gnupg openssl && \
+    gpgconf --kill dirmngr && \
+    gpgconf --kill gpg-agent && \
     rm -rf /root/.gnupg
+
+COPY tar /bin/
+
+COPY docker-entrypoint.sh /bin/docker-entrypoint.sh
+
+# Lunacluent install
+
+## set up luna client tar file
+#COPY <610-000401-003_SW_Linux_Luna_Minimal_Client_V10.3.0_RevA.tar> /tmp/
+
+#RUN mkdir -p /usr/safenet/lunaclient
+#RUN mkdir -p /usr/safenet/lunaclient/cert
+#RUN mkdir -p /usr/safenet/lunaclient/cert/client
+#RUN mkdir -p /usr/safenet/lunaclient/cert/server
+
+## set up luna client tar file
+#RUN tar -xvf /tmp/<610-000401-003_SW_Linux_Luna_Minimal_Client_V10.3.0_RevA.tar> --strip 1 -C /usr/safenet/lunaclient
+#RUN cp /usr/safenet/lunaclient/openssl.cnf /usr/safenet/lunaclient/bin
+
+#ENV ChrystokiConfigurationPath=/etc
+
+
+##Set up cert, Chrystoki.conf 
+#ㅈCOPY Chrystoki.conf /etc/Chrystoki.conf
+#COPY 10.22.20.32.pem /usr/safenet/lunaclient/cert/server
+#COPY 10.22.20.32Key.pem /usr/safenet/lunaclient/cert/client
+#COPY server.pem /usr/safenet/lunaclient/cert/client
+
+
+
+## add vault group user
+RUN groupadd --gid 1000 vault && \
+    adduser --uid 100 --system -g vault vault && \
+    usermod -a -G root vault
+
+
 
 # /vault/logs is made available to use as a location to store audit logs, if
 # desired; /vault/file is made available to use as a location with the file
 # storage backend, if desired; the server will be started with /vault/config as
 # the configuration directory so you can add additional config files in that
 # location.
+ENV HOME /home/vault
 RUN mkdir -p /vault/logs && \
     mkdir -p /vault/file && \
     mkdir -p /vault/config && \
-    chown -R vault:vault /vault
+    mkdir -p $HOME && \
+    chown -R vault /vault && chown -R vault $HOME && \
+    chgrp -R 0 $HOME && chmod -R g+rwX $HOME && \
+    chgrp -R 0 /vault && chmod -R g+rwX /vault
 
 # Expose the logs directory as a volume since there's potentially long-running
 # state in there
@@ -47,3 +108,12 @@ VOLUME /vault/file
 # 8200/tcp is the primary interface that applications use to interact with
 # Vault.
 EXPOSE 8200
+
+COPY docker-entrypoint.sh /usr/local/bin/docker-entrypoint.sh
+ENTRYPOINT ["docker-entrypoint.sh"]
+
+USER 100
+
+# By default you'll get a single-node development server that stores everything
+# in RAM and bootstraps itself. Don't use this configuration for production.
+# CMD ["server", "-dev"]
